pyproject.toml
[tool.poetry]
name = "lab_eng_sw"
version = "0.1.0"
description = ""
authors = ["taciossbr <tacioss.br@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django-rest-framework = "^0.1.0"
requests = "^2.22"
coreapi = "^2.3"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
